+++ tests/utils.mjs
@@ -177,6 +177,7 @@ function parseTestFile(tob) {
   tob.shouldPassModule = tob.aboveTheFold.toLowerCase().includes('\n## pass module\n');
   tob.shouldPassSloppy = tob.aboveTheFold.toLowerCase().includes('\n## pass sloppy\n');
   tob.shouldPassAnnexb = tob.aboveTheFold.toLowerCase().includes('\n## pass annexb\n');
+  tob.shouldPassAnnexboth = tob.aboveTheFold.toLowerCase().includes('\n## pass annexboth\n');
   tob.shouldPassAny = tob.aboveTheFold.toLowerCase().includes('\n## pass\n') && !tob.shouldPassModule && !tob.shouldPassSloppy;
   tob.shouldFail = tob.aboveTheFold.toLowerCase().includes('\n## fail\n');
 
@@ -287,7 +288,7 @@ function parseTestFile(tob) {
       return obj;
     }, {});
 
-  const {es, astUids, locationTracking, ranges, nodeRange, exposeScopes, alwaysAllowOctalEscapes, ...unsupported} = tob.inputOptions
+  const {es, astUids, locationTracking, ranges, nodeRange, exposeScopes, alwaysAllowOctalEscapes, allowUsingDeclaration, ...unsupported} = tob.inputOptions
 
   // If this triggers then the line above may need updating
   ASSERT(JSON.stringify(unsupported) === '{}', 'options have hardcoded support in the test suite so if a new option needs support, make sure to connect it first, then update this assert. Unhandled options: ' + JSON.stringify(unsupported) + ', file: ' + tob.fileShort);